LinuxHeaders: ApplyKernelConfig()
Summary:
This diff finds a kernel config file on the host machine, and uses it to generate
an autoconf.h to patch into the prepackaged linux kernel headers.

This is required so the right define statements are used.

It is also a blocker for AWS/EKS.

Test Plan: Added a test case for GenAutoConf(). Also manually tested on EC2.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1944

Differential Revision: https://phab.corp.pixielabs.ai/D4922

GitOrigin-RevId: 350933136080fc63b3571ff0a6cfc8dfec39abeb