PC-283: Add graphql handlers for listing all clusters, and looking up a cluster by ID
Summary: This is to support PC-283 which will add an admin UI page. We need to be able to list all clusters a user has, and also look up a cluster and its info by ID. Currently we just take the first cluster, and that way will be deprecated once the UI is updated to use the new endpoints.

Test Plan: added/existing

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-283

Differential Revision: https://phab.corp.pixielabs.ai/D4948

GitOrigin-RevId: 98a157ca20a7b332692c175d4a8b1e018a90d750