Add bootstrap option to VizierHeartbeat
Summary:
we want to support pushing updates/deploys from the cloud. to do so, cloudconn reads from a configmap with BOOTSTRAP=true, and the version it should be deploying to.
it then sends this message to vzmgr, and vzmgr triggers a deploy/update if the value is true. if it receives a heartbeat with BOOTSTRAP=true and the vizier cluster is already updating, it should not start another update.
originally i wanted this to be in the vzconnected message instead, so that it is only sent once. but there is a race condition there where cloudconn only starts listening to NATS messages through vzconn after VizierConnected has returned successfully.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4990

GitOrigin-RevId: 014b04a9bd2883f010a19ea55afe72eed37cd82c