Cancel the grpc request on the client side.
Summary: Instead of setting the deadline header in the GRPC request (it meant the stream has to complete by the deadline, which is not actually what we want), use a timeout mechanism in the UI and cancel the request if timed out.

Test Plan:
tested manually
For the healthcheck case I disabled passthrough and the request gets cancelled,
for the execute script case I introduced some network delay and the request errors as expected.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-297

Differential Revision: https://phab.corp.pixielabs.ai/D4991

GitOrigin-RevId: a79d0a5afde388f12849a275018bd173e8ea7aa5