Allow creating augmented tokens from non-user tokens
Summary:
for updating our deploy process, we want to be able to issue a cluster token to the user's vizier so that it can make a call to download artifacts.
since the deploy is not triggered by a user, it is not linked to a user token.
instead, we want to generate a cluster token for the cluster. that means we should allow non-user tokens in GetAugmentedToken.
NOTE: some calls will still need a user token (it gets the user id/org id). if these calls are given a service or cluster token, they will fail. (as expected). calls that are not user dependent, like downloading artifacts, should work fine.

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4988

GitOrigin-RevId: c467c9471aa4e315d35dd395357fa7074cc09012