Update vzmgr to initiate deploy on bootstrap mode
Summary:
vzmgr needs to initiate a deploy when it sees that a vizier is in bootstrap mode.
if the vizier is already updating, it should not launch another update.

the logic here is similar to the one that already exists for the update path:
- if no version is specified, fetch the latest
- check that the specified version is valid
- send a message to the cluster
the main difference is that we use a cluster JWT instead of a user JWT, since the update was not initated by a user

Test Plan: tested on separate branch, and also checked that update path still works.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4995

GitOrigin-RevId: ae76b578618b4ac3674bbf0e6a77f70bf7b37b8e