Update cloudconn to send bootstrap mode to vzmgr
Summary:
we want to trigger deploys/updates from the cloud. to do so, if cloudconn starts up with a bootstrap config where the value is true, then it should send that config to vzmgr which will trigger the update/deploy.
this diff just updates cloudconn to read the config and send it to vzmgr.

the followup diff will actually update the CLI deploy to create this bootstrap configmap

Test Plan: ran in separate branch

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4996

GitOrigin-RevId: 25b30bca5fdc68f1c471da8c7600862562dc918b