Only update the browser history when a script is executed.
Summary: Depends on D4968. Before this change, any update to the arguments not only updates the URL, but it also creates a new history entry. This avoids polluting the browser history by calling replaceState instead of pushState, and pushState when a script is executed.

Test Plan: updated tests. tested manually

Reviewers: michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D5005

GitOrigin-RevId: 03bc5fa5a087455f1520f1589dc9297a348e7310