PC-292: Vizier heartbeat should send the current number of nodes in the k8s cluster
Summary: tsia, this is part of PC-220. The goal is for the admin page to show the % of nodes on a cluster that are instrumented by vizier. We have the total number of agents already, and so this part is used to know the total number of nodes on the cluster.

Test Plan: open to adding a test but it looks like the heartbeats aren't tested right now

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-292

Differential Revision: https://phab.corp.pixielabs.ai/D5014

GitOrigin-RevId: 670925f9ca85029240b7c169f049bfdc947d4112