Fix spammy K8s API errors in metadata
Summary:
we were seeing extremely spammy messages in the metadata service along the lines of "Received non-namespace object when handling namespace metadata."
After some debugging, it looks like this occurs on clusters where there have been no namespace updates for a long time. this is what happens:
1. We start up metadata service and do and initial sync to etcd up to date/delete resources that no longer exist. We record the last resourceVersion for each resource.
2. We start up the K8s watcher to start watching from the last resourceVersion. This is so we don't see duplicate events.
3. This is where the bug occurs: for resources which haven't been updated for a long time, their last resourceVersion is very old. Plugging this number into the k8s watcher results in an error. the k8s watcher mass-spams nil objects to the channel, which is what we were seeing.
The fix:
we get the latest resourceVersion from all of our resources and start watching from that resourceVersion. this resourceVersion should not be too old, and is likely the resourceVersion for some pod.

Test Plan: ran skaffold, hardcoded a low RV for namespaces, and checked to make sure that the k8s watcher restarted and ran fine after the initial failure

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5026

GitOrigin-RevId: cdb32e0eb8cec731abf9cd3dace3a94d88a33503