Update CLI Deploy to support extract YAMLs
Summary:
this involved refactoring all of k8s resource creation to instead generate a YAML.
if the extract_yaml option is specified, then we write the YAMLs to a tar file.
otherwise, we deploy the YAMLs

Test Plan:
deployed a bunch of times, and also deployed using the extracted yamls.
there does appear to be some flakiness with cloudconn receiving the update/deploy request on staging (but not on plc-dev...) but after some investigation, i am running into this issue with the old deploy as well.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5013

GitOrigin-RevId: 310135d31445abd5aaf9576010e3df5a391ce612