Add --redeploy_etcd flag to px update vizier
Summary:
we are having to redeploy etcd very often. to make the process easier we want to add a --redeploy_etcd flag to indicate that etcd should be restarted on the update.
we are modifying the clusterrole, so I just updated the clusterrole with a new name. this way, any newly installed clusters will already have the correct updaterrole permissions. an alternative to this is to update the updaterjob to use the k8s API to update the clusterrole, but this seemed a little uglier to me since its really only a one-time operation.

your vizier must be updated to this version in order for --redeploy_etcd to work, because the vizier needs the redeploy_etcd field in the proto.

Test Plan: tested running update job (with flag always set to true). can't actually test the whole flow on plc-dev, because updates don't work on cloud dev.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5032

GitOrigin-RevId: bfaaac6578075828f85460229da1a051ac2d8592