ProvisionVizier assigns clustername if specified & current name is null
Summary:
if ProvisionVizier is given a recommended cluster name, it should use the clusterName if there is none already specified.
turns out the current logic for retrying a name until one does not collide doesnt work. as soon as the first collision returns an error, the entire transaction is canceled.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5120

GitOrigin-RevId: 61df59c494a244d0407464a58fa7066a21be3b3d