Keep the top bar in the loading screen.
Summary:
Problem: When the user tries to switch to a different cluster, the UI switches to the Connecting screen, and there is not a whole lot they can do other than to wait, and if the cluster is unhealth or there are connection issues, they are stuck forever.

Solution: Move the loading screen to be inside the live view, so the top bar is still available for switching to a different cluster or navigating to the admin view.

Test Plan: tested manually

Reviewers: michelle, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5127

GitOrigin-RevId: 317aca73362f4bd6db4ebf32e0a26cbc3800697a