SemTypes Pt. 1: Add objects to represent types in the compiler.
Summary:
All Operators and Expressions in the compiler will now have a Type that is a subclass of BaseType. Operators (at least for now) will all have TableTypes since all operator's currently operate on columnar data.
All Expressions currently have ValueType, since we don't allow non-scalar expressions at the moment in the IR. In a future PR, these types will support hashing so that type checking can be done more easily.

Also, this begins the process of decoupling the compiler's handling of types from the Relation object (which will help in the future if we need to change things because of tags)

Test Plan: Added tests for the type objects themselves, tested more with future diffs.

Reviewers: philkuz, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5217

GitOrigin-RevId: 1bf1429d9dbb2dc363d66afbb31e23caa7e8cf39