SemTypes Pt. 2: Add semantic type inference rules for UDFs.
Summary:
This diff adds support for specifying how UDFs propagate semantic types.

`type_inference.h` adds subclasses of InferenceRule that can be used to specify how the UDF should propagate semantic types.
For instance, the UDF can explicitly specify the output type for a given list of input types using the ExplicitRule, or it can specify that for certain semantic types, the output type should be whatever the input type of certain arguments are.

This diff adds a couple examples of the usage of these rules in AddUDF and MeanUDA. A later diff adds more comprehensive rules.

Test Plan: Added tests that the udf registry handles the rules properly, and that RegistryInfo can deal with them too.

Reviewers: philkuz, nserrino, #engineering

Reviewed By: nserrino, #engineering

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D5218

GitOrigin-RevId: 429ae0ac9d6a0b7c45edbd66881784bb873fbf98