PC-327: Fix white screen when '{}' is provided as vis spec
Summary:
We need to refactor parseVis to do better enforcement/type checking of the provided vis.json.
That will happen when we enforce the json schema of the vis protobuf. For now, let's prevent common white screens
in the live view ui with this simple fix.

Test Plan: ran it with {} as vis spec.

Reviewers: philkuz, malthus, michelle, #engineering

Reviewed By: malthus, #engineering

JIRA Issues: PC-327

Differential Revision: https://phab.corp.pixielabs.ai/D5158

GitOrigin-RevId: de8e7e3cbbba1e34f9d4e70f369b54c504f28787