Mark data/control plane in vizier YAMLs
Summary:
we want to distinguish between whether a pod is in the control/data plane for running vizier. we add this information as a label in the vizier YAMLs.
cloudconn can then use this label to filter out the data plane pods (kelvin/pem) when sending the pod statuses in its heartbeats

Test Plan: ran skaffold, printed out the pods in cloud conn and check that pems/kelvin is filtered out

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5157

GitOrigin-RevId: 8cc4103bb7ac747179411a339d17e78640b20b3d