Use response time as the time column
Summary:
This is required because for protocols that support multiple parallel requests, there is no way for the data table to keep requests sorted by request time.

In the worst case, it would have to wait indefinitely before pushing out a record. This is particularly bad, since we operate in TransferData() iterations.

Test Plan: Added a test case.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1986

Differential Revision: https://phab.corp.pixielabs.ai/D5141

GitOrigin-RevId: e721233f6bdc4709760251f869c8c8a2a001d57a