PC-306: Add (barebones) entity landing page for cluster
Summary:
Similar to the other diffs out like this. This entity page for /cluster/:clustername is the most barebones yet.
I am open to adding other stuff, but for now it's a placeholder that mostly servces as a jumping off point to click on
hyperlinks into the more interesting resources (such as namespace or node).

Test Plan: ran it in the UI

Reviewers: philkuz, zasgar, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PC-306

Differential Revision: https://phab.corp.pixielabs.ai/D5150

GitOrigin-RevId: ce03447837a9d6b0b4e4d17b4afe0010d9139deb