Make artifacts gql request unauthenticated
Summary:
we now allowe unauthenticated users to view the docs page. that means that we need to make the artifacts request unautheticated.
to do so, I added a path for unauthenticated gql requests.

Test Plan: ran cloud dev and docs. checked that docs could access the artifacts and that other cloud gql requests (such as getClusters) still required auth and still worked.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5172

GitOrigin-RevId: 131098e23ca7715cae93fb82e330a9e6c0c811bc