PC-306: Add entity pages for namespace and namespaces
Summary:
These will go in /cluster/:clustername/namespaces and /cluster/:clustername/namespace.
They are a bit sparse because I think that most of the other items we could put here are going to go in
/services, /pods, etc. We can choose to duplicate if we want but I thought that it made sense for these views
to be a bit more of a jumping off point, where the user clicks a hyperlink or visits the /pods entity page.
However, open to feedback. Screenshots will be attached.

I was kind of surprised that I got so little IO in my namespace summary, so if someone sees the error in my
logic there pelase let me know. I will also double check it some more.

Test Plan: ran the UI

Reviewers: philkuz, zasgar, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

Subscribers: ishan

JIRA Issues: PC-306

Differential Revision: https://phab.corp.pixielabs.ai/D5149

GitOrigin-RevId: 06fd5e9dab2298a7aa73075140d313053b37b1dc