Rewrite data drawer table to be more performant
Summary:
The original data-table used to measure each table cell to better distribute column widths. However it had to measure every row of data before it can render so it frozed up the UI on large datasets. This diff changes the table to use flexbox to distribute the column width, and it sets the flex basis based on the content of each row.

The main difference between with the new data table are:
- Tables now always span full width of the container
- Tables can no longer scroll horizontally
- Mouse hover highlights the whole row instead of individual cells

Test Plan: Tested manually with most existing scripts. data drawer now render in 25ms vs 10 seconds with ~10k records

Reviewers: michelle, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-231

Differential Revision: https://phab.corp.pixielabs.ai/D5208

GitOrigin-RevId: 89ff0c81aa0c3955181a679e065e2f2bcc866ee8