Rename vzmgrpb package -> pl.services.internal
Summary:
prototool is getting confused because we have two proto messages in the same package with the same name. we'll run into more instances of this in the future, so simply renaming
one of the messages doesn't seem like a good solution. instead, we can rename the internally-used proto from pl.services -> pl.services.internal. this way, there will be no collisions.

Test Plan: ran plc-dev. things seemed to work as normal

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5212

GitOrigin-RevId: 3464703ed647730cb464968abea098e9dd4a28ff