PC-304: Add entity centric urls for pods, services, nodes.
Summary:
Depends on D5237. Adds the routes for the above entity live views.
doesn't do the cluster one yet because it's unclear how the url for that should difffer from the generic live view url.
Note that some special handling of pod and service is required because in Carnot we prepend namespace to those names,
but we don't in our entity url scheme.

Test Plan: added/existing/ran the various urls

Reviewers: malthus, michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-304

Differential Revision: https://phab.corp.pixielabs.ai/D5241

GitOrigin-RevId: dc1a58029089c8de14c795b9ff5a637fd99a44f9