PP-2011 Deploy vizier deps when updater job in bootstrap mode
Summary: there is a race condition when deploying with YAMLS where we try to deploy the nats cluster before the crd is created. there doesn't seem to be a way to add a wait using just the YAMLs, so we need to deploy the deps through the updater job.

Test Plan: deployed with/without yamls, making sure to delete nats crd

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5263

GitOrigin-RevId: 34ddf9a7669ffbd67602ac06e29868f6194a50dd