Fix px deploy for plc-dev
Summary:
px deploy was broken because the update job was trying to use "vzconn-service.plc-dev.svc.cluster.local:51600" as the API service cloud addr.
we should instead be using "api-service.plc-dev.svc.cluster.local:51200" as the cloud addr in the update job.

an alternative to using a configmap variable is to check in the update job if the cloud addr contains vzconn, and use the api-service addr instead. but this seemed a little cleaner.

Test Plan: ran px deploy with plc-dev, checked that vizier starts up and running `px get pems` works

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5265

GitOrigin-RevId: 118648ba00aa42ee067731db833f78031da25976