Fix "reindexing not supported bug" with log collector
Summary:
There was an issue with the ManagedIndex where on rollover sometimes the mapping wouldn't be updated with the mappings from the old index, and then on the first write the mapping would be set dynamically, then when the log collector is redeployed, it notices that its mapping is inconsistent with the mapping of the new index and so it would throw, saying that reindexing is not supported.

Its unclear to me why elasticsearch doesn't always maintain the mappings of indices on rollover. But by adding the mapping to the Index Template it should force elastic to include the mappings in the rollover indices.

Test Plan: Ran skaffold a bunch of times and waited for rollovers and then ran skaffold, and had no issues.

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5266

GitOrigin-RevId: 0faa50b93be798779e511a549fb8e3bc7f53dde2