GenVariable() to generate code for a Variable
Summary:
Contrary to what I stated in the meeting, message Variable is adjusted to describe Register
or a variable in memory.

The reason is to make it natural for variable that needs to base off SP pointer.
Otherwise, for a variable base off SP with 8 bytes, intermediate IR -> physical IR needs to insert
the base as an imaginary SP variable in the BCC code. In other words, an implementation detail of
the physical IR -> BCC code generation is leaked into intermediate IR -> physical IR transformation.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-2008

Differential Revision: https://phab.corp.pixielabs.ai/D5289

GitOrigin-RevId: 8fb20e73996fe8aa1d90a0441592dcae45fe04c1