Fail if cloudconn can't connect to NATS when not in bootstrap mode
Summary:
we updated cloudconn to not fail if it can't connect to NATS. this is so we can provision NATS later in the update job.
however, it's possible that NATS may take some time to start up. when the updater job relaunches cloudconn, NATS may not be ready yet. with the current behavior, it won't try to reconnect to nats and keep using a faulty nats connection.
instead, if cloudconn is not in bootstrap mode, we should expect it to be able to connect to NATS. this way cloudconn will restart until NATS is ready.
if this fixes the bug, we should update cloudconn to try to connect to nats with an exponential retry rather than just doing a Fatal.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5304

GitOrigin-RevId: 60cca514d566744c1f58e8c34795151d780e2790