Fix text alignment in the table
Summary: The expansion icon was causing the items to long align correctly. This fixes the issue. It might be a better design for us to actually add a gutter column to the table for these actions. We can investigate that as we cleanup the table implementation.

Test Plan: Tested on webpack dev server.

Reviewers: michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5314

GitOrigin-RevId: 6ccb441835072813a6baed55a4e8cce9c23a2baa