PC-380 Sort histogram axis by value
Summary: Depends on D5323. This undoes a previous change that blanket changes the default renderer to use the string renderer without adjusting who actually needs the identity renderer. Made it so canvas uses the identity renderer.

Test Plan: Tested locally and histogram works again. This is more of a reversion so we expect it to work.

Reviewers: jamesbartlett, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-380

Differential Revision: https://phab.corp.pixielabs.ai/D5332

GitOrigin-RevId: 46676554583b95378bd9f0c0df3e5e67d7f3e00b