Add GenMapStashAction()
Summary:
StashMapAction now only have one variable_name, and it should refer to one of the predefined args,
ret_vals, function_latecy in logical & intermediate IR, or one of the predefined vars
in the physical IR.

That leaves the variable name generation to the intermediate -> physical IR translation,
which should be preferred as it allows BCC code gen to be literal translation.

Rename ValueType to VariableType, as it describes variable not value.

Rename StashMapAction to MapStashAction.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-2008

Differential Revision: https://phab.corp.pixielabs.ai/D5321

GitOrigin-RevId: d3ce339fb48a70f527ab68448633b1a7f38ee55f