Don't render the data until we actually show it
Summary:
This is a constant source of bugs because we keep accidently using the rendered values.
This changes everything to be stored in native pre-rendered format until we actually render the tables.

The canvas tables have not been updated yet, but will be in a future diff.

Test Plan: Tested locally.

Reviewers: michelle, nserrino, philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5359

GitOrigin-RevId: ddf02d6b94bd2934fcc146f136c72a27d6425c84