Add new autocomplete input component
Summary:
This is the new autocomplete input component, a modified version of autocomplete/input.tsx.  When the old autocomplete is removed from the app, we can delete those files and rename these.
this component is mainly necessary because because the standard HTMLElement input doesn't allow us to manipulate textColor of individual characters.
the input component is responsible for rendering the text in what looks like an input box, and for listening to click/keyboard events like a normal input box does. this is actually built by layering a regular div ontop of a standard HTMLElement input with opacity 0.

Test Plan: jest test

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5386

GitOrigin-RevId: e8c47807672f1fa1c382c8ff8086ca23b1e43128