PC-397: Fix UX issue where opening editor causes the widgets to be stuck in 'moveable' state
Summary:
Editor is now openable through the side caret which is always there except in mobile.
Icon at the top is now move widgets icon, which disappears in mobile view.
We should also add an escape key shortcut to turn off the move widgets mode, because it is a common interaction pattern and because I don't know if it is super obvious to click the move widget icon to escape.

add eslint

Test Plan: ran it in various cases including mobile

Reviewers: michelle, philkuz, zasgar, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-397

Differential Revision: https://phab.corp.pixielabs.ai/D5402

GitOrigin-RevId: c1621f8d323cbaaa682604e03ece87d9315f9cb5