UI: Add autocomplete to args editor
Summary:
this is basically Malthus's work, with some cleanup and testing.
similarly with the CLI, we try to determine the type of the arg (svc/pod/ns) from the argument name. in the future when we index our scripts + their arg types in elastic, we won't need to do this.
if we've detected that the argument is a known entity, then we can submit an autocomplete request to the backend and show a list of suggestions.

since the backend isn't ready yet, this operates mostly like the current argument editor.

Test Plan: ran webpack

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5307

GitOrigin-RevId: 55baa26428c56b409a9a0f598dff41eebbd38a8a