PC-414: Make pixie live view page title show up as the entry in the browser history
Summary:
This change also refactored title a bit. title is a script context property that comes from the script.
It is always either the title for a certain script id, or the name of the entity. so we don't actually have to have
execute() set or care about it at all, because it can be deterministically produced from id + entity in script context.

Test Plan: ran it in various cases. this also exposed PC-418.

Reviewers: zasgar, philkuz, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-414

Differential Revision: https://phab.corp.pixielabs.ai/D5406

GitOrigin-RevId: 7fc6201af40b1a55f585a6eeb44441d38533bbbe