PC-421: quantiles should be a semantic type, remove latency box whisker from vis.proto.
Summary: After doing a few iterations on the quantiles in the frontend, it turns out to be scalable and cleaner to have quantiles be a semantic type in pixie rather than something that gets annotated on at the end. The frontend expects certain values for quantiles plots (p50, etc) and this semantic type will provide a better guarantee that a given column matches that format.

Test Plan: n/a

Reviewers: jamesbartlett, philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-421

Differential Revision: https://phab.corp.pixielabs.ai/D5442

GitOrigin-RevId: 620949542fa5ca9df3d1347667378770685b1422