Tweak arc proto unit tests
Summary:
Stirling uses proto files internally, but the arc unit tests are adding a bit too much friction to our development. So a few tweaks:

1) Add stirling/dynamic_tracing/ir to list of prototool excludes. These protos are not public, so should be safe for us to introduce "breaking" changes.
2) Remove the check for the initial pb.go check-in.
Rationale: It should be up to the developer to decide whether they want a pb.go file.
In our case, we don't generate them anymore.
Note that once the pb.go file is checked in, the checker still ensures that it remains up-to-date from that point on.
So the only check that is removed is the initial pb.go check-in...and if the pb.go file was really required, then something else would break in the diff if someone forgot to check it in. So we should be okay.

Test Plan: Jenkins

Reviewers: michelle, #engineering, yzhao

Reviewed By: michelle, #engineering, yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D5445

GitOrigin-RevId: a5f1af34c1a5ec9f1e636ff953fcbecc2f288317