PC-422: Add quantiles box whisker plot to components.
Summary:
This is to support rendering latency quantiles as a box whisker.
I went back and forth on this a few times but decided to make it a generic quantiles component, not latency-specific.
It does take a level low/med/high to decide the color, which is a bit latency-specific where high=red. however,
this is a minor change if we want to use other colors for other values down the line (also the use of a level to
cause a certain color text to show up is not required).

Test Plan: ran it, tests will go around the logic itself

Reviewers: philkuz, jamesbartlett, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PC-422

Differential Revision: https://phab.corp.pixielabs.ai/D5441

GitOrigin-RevId: cf64ee88044cfa31eb165ed7616a4c501f1a2728