CLI: Move YAML generation to pkg
Summary:
deploy.go is currently a mess. moving yaml generation/extraction to a separate file for cleanliness.
this refactor doesnt yet include templating YAMLs.

the idea for the near future is that generateYAMLs will generate templateable yamls. calling ExtractYAMLs or any of the Get___YAML functions with the template arguments will fill in the template.

Test Plan: ran deploy with/without extract YAMLS

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5452

GitOrigin-RevId: dc50da2aeb0ec54be932f8b27222ff9f1fb343aa