Simplify the way args are stored in the UI.
Summary: The prior way args were stored (separately as entity args and non entity args) was a bit of a holdover/temporary state from the refactor to supporting entity pages. Now that that refactor is mostly complete, we can streamline the way that args are stored and simplify it a bit. Now entity param names are gotten from a map of LiveViewPage to the set of args that are specific to that entity. Args are stored all in the same place, and when entity-specific or non-entity rags are needed, they are produced on the fly.

Test Plan: ran it in various cases

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5431

GitOrigin-RevId: e4c1b84feaba9dfdc06fa269055390f8ddb9610a