PC-420 - Add pretty cluster name to the GRPC API
Summary:
We use hueristics to extract a more user friendly and shorter cluster name.

Right now this name is just generated in the API layer. Eventually we probably want to move this to VZMGR and allow this name to be customized.

Test Plan: Added tests for naming function and API.

Reviewers: michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-420

Differential Revision: https://phab.corp.pixielabs.ai/D5449

GitOrigin-RevId: 71d9069b01b8b803500790bb222e03fe7d2a265f