Templatize deployKey in deployment yamls
Summary:
the helm chart needs the deployKey to be templatized. updated the extractYAML fn to take tmpl arguments, for if the template needs to be filled out.
using the template execution in go, this works correctly.

Not that the templated YAML does contain strings around the template value itself:
```
stringData:
  deploy-key: '{{ .Values.DeployKey }}'
```

Test Plan: deployed with/without extract yamls

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5463

GitOrigin-RevId: 3a626a4b563a344e4c449a367048827616122517