Process Logical probes
Summary:
This diff creates and entry and return probe for a given logical probe.

It also creates the required maps and output buffers.

Currently, no optimizations are performed to remove any unused probes.
For example, if only an entry probe is required, this transformation will still create a return probe,
and transfer the args to the return probe, which will then perform the push.

Test Plan: Test case added.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5454

GitOrigin-RevId: dfec45a0d925413ac0285bab37e1e891b355a239