Add fixedSizeDrawer component
Summary:
we want to use the material UI drawer component. to do that, we need a wrapper around the drawer content + the content outside of the drawer so that the outer content shrinks properly when the drawer is opened/closed. this is because MUI gives the drawer a fixed position, and takes it out of the DOM.  this is following the persistent drawer example in the MUI documentation.
to allow for resizing, there needs to be another wrapper around this component which passes in the correct width/height. that will come in a following diff.

Test Plan:
replaced data drawer to use fixedSizeDrawer to make sure it behaves as expected. this is coming in a followup diff.
there is a version in storybook, but it doesn't look very good since the drawer is docked on the top/bottom/left/right of the entire storybook page.

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5484

GitOrigin-RevId: d98f4889ae0c7b3b33796953f162ea831611b6f1