Add resizable drawer component
Summary: we should be able to resize our drawers. this is done using react draggable, and we pass the calculated width/height to the fixedSizeDrawer.

Test Plan: updated data drawer to use resizable drawer. the storybook is bit weird because the drawer is docked outside of the storybook container, so the resizable handle is not in the expected position.

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Subscribers: nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D5487

GitOrigin-RevId: 002cb75ef1107ba4a2e13d1786f39da59111a32f