PC-429 Render text for missing table instead of dropping a widget.
Summary:
Had a reallly annoying interaction where charts would be dropped if the backing table was missing. This meant if you had a typo in a function name, your chart would be dropped. This would happen with global functions missing (see PC-429 for a repro).

Need to just verify that the vis spec is written correctly before we execute, doing in a follow up diff. Just adding this because this is a simple fix that improves experience significantly.

Test Plan: Tested locally and does not nuke the vis spec if you miss a function.

Reviewers: jamesbartlett, michelle, zasgar, #engineering

Reviewed By: jamesbartlett, #engineering

JIRA Issues: PC-429

Differential Revision: https://phab.corp.pixielabs.ai/D5497

GitOrigin-RevId: 0b6a84dc793146dba2044f30865a5adb79415c55