[PC-429] Add validation for visSpec
Summary:
Before, visSpec would be unvalidated when we run it, so you'd have weird side-effects that are hard to explain. This adds validation that explains why something is broken and can lead to eventually directing users to where / how to update scripts.

This improves the experience a bit.

Test Plan: Tested in yarn dev and also added new tests for the validation code.

Reviewers: jamesbartlett, nserrino, michelle, #engineering

Reviewed By: jamesbartlett, #engineering

JIRA Issues: PC-429

Differential Revision: https://phab.corp.pixielabs.ai/D5500

GitOrigin-RevId: 6a8a4cd031e9e9b489ff18421973716008a5b13c