Metadata: better handle nil hostname/IP pairs
Summary:
customer is running into an issue where there is a nil hostnameIPPair when we are trying to resolve an endpoints update to the relevant agent.
tracking the hostnameIPPair through the code, it looks like it is nil because we can't find the hostnameIPPair for a given pod in etcd...
unfortunately its hard to say why this info is not in etcd. for now, we can handle it better so that we don't crash, and also log the pod name to see if there's anything unique about this pod.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5503

GitOrigin-RevId: de1e92a5ab7c7e519dd90aa1b36d88535a31904b